Refactor + Setup for testing
