+++ components/ProductItem/ProductItem.test.js
@@ -1,0 +1,5 @@
+describe("yay", () => {
+  it("works", () => {
+    expect(2).toEqual(2);
+  });
+});
