+++ package.json
@@ -15,9 +15,9 @@
   },
   "devDependencies": {
     "@babel/core": "~7.9.0",
-    "babel-jest": "~25.2.6",
-    "expo": "^38.0.10",
+    "@testing-library/react-native": "^7.0.2",
     "babel-eslint": "^10.0.3",
+    "babel-jest": "~25.2.6",
     "eslint": "^6.8.0",
     "eslint-config-airbnb": "^18.0.1",
     "eslint-config-prettier": "^6.9.0",
@@ -26,6 +26,7 @@
     "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-react": "^7.17.0",
     "eslint-plugin-react-hooks": "^2.3.0",
+    "expo": "^38.0.10",
     "jest": "~25.2.6",
     "metro-react-native-babel-preset": "^0.57.0",
     "prettier": "^1.19.1",
